dsa-backend/pyproject.toml
[project]
name = "dsa-back"
version = "0.1.0"
description = "DSA Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=44.0.2",
    "fastapi>=0.115.11",
    "httpx>=0.28.1",
    "jsonschema>=4.23.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.1",
    "pytest>=8.3.5",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "pytz>=2025.1",
    "sqlalchemy>=2.0.38",
    "uvicorn[standard]>=0.34.0",
    "xlrd>=2.0.1",
]
